update printpdf to 0.6.0 and remove transparency restriction

+++ src/fonts.rs
@@ -353,9 +353,9 @@ impl Font {
             idx,
             is_builtin,
             scale,
-            line_height: printpdf::Pt(f64::from(line_height)).into(),
-            glyph_height: printpdf::Pt(f64::from(glyph_height)).into(),
-            ascent: printpdf::Pt(f64::from(ascent)).into(),
+            line_height: printpdf::Pt(f32::from(line_height)).into(),
+            glyph_height: printpdf::Pt(f32::from(glyph_height)).into(),
+            ascent: printpdf::Pt(f32::from(ascent)).into(),
         }
     }
 
@@ -386,7 +386,7 @@ impl Font {
     /// [`FontCache`]: struct.FontCache.html
     pub fn char_width(&self, font_cache: &FontCache, c: char, font_size: u8) -> Mm {
         let advance_width = self.char_h_metrics(font_cache, c).advance_width;
-        Mm::from(printpdf::Pt(f64::from(
+        Mm::from(printpdf::Pt(f32::from(
             advance_width * f32::from(font_size),
         )))
     }
@@ -399,7 +399,7 @@ impl Font {
     /// [`FontCache`]: struct.FontCache.html
     pub fn char_left_side_bearing(&self, font_cache: &FontCache, c: char, font_size: u8) -> Mm {
         let left_side_bearing = self.char_h_metrics(font_cache, c).left_side_bearing;
-        Mm::from(printpdf::Pt(f64::from(
+        Mm::from(printpdf::Pt(f32::from(
             left_side_bearing * f32::from(font_size),
         )))
     }
@@ -422,13 +422,13 @@ impl Font {
             .get_rt_font(*self)
             .glyphs_for(s.chars())
             .map(|g| g.scaled(self.scale).h_metrics().advance_width)
-            .map(|w| Mm::from(printpdf::Pt(f64::from(w * f32::from(font_size)))))
+            .map(|w| Mm::from(printpdf::Pt(f32::from(w * f32::from(font_size)))))
             .sum();
         let kerning_width: Mm = self
             .kerning(font_cache, s.chars())
             .into_iter()
             .map(|val| val * f32::from(font_size))
-            .map(|val| Mm::from(printpdf::Pt(f64::from(val))))
+            .map(|val| Mm::from(printpdf::Pt(f32::from(val))))
             .sum();
         str_width + kerning_width
     }
